1.0.0

+++ src/main/java/net/nova/nmt/event/NMTEventBus.java
@@ -23,7 +23,6 @@ import net.minecraft.world.level.gameevent.GameEvent;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.neoforged.neoforge.registries.datamaps.builtin.Compostable;
 import net.nova.nmt.init.NMTItems;
 import net.nova.nmt.init.NMTPotions;
 
@@ -75,7 +74,7 @@ public class NMTEventBus {
                                     serverlevel.sendParticles(
                                             ParticleTypes.LAVA,
                                             (double) blockpos.getX() + serverlevel.random.nextDouble(),
-                                            (double) (blockpos.getY() + 1),
+                                            blockpos.getY() + 1,
                                             (double) blockpos.getZ() + serverlevel.random.nextDouble(),
                                             1,
                                             0.0,
